big refactor

+++ order-service/src/main/java/com/example/orderservice/repository/OrderRepository.java
@@ -1,9 +1,9 @@
 package com.example.orderservice.repository;
 
-import com.example.orderservice.model.Order;
+import com.example.orderservice.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Long> {
+public interface OrderRepository extends JpaRepository<Order, Integer> {
 }
