+++ shipping-service/src/main/java/com/example/shippingservice/repository/ShippingRepository.java
@@ -1,9 +1,9 @@
 package com.example.shippingservice.repository;
 
-import com.example.shippingservice.model.Shipping;
+import com.example.shippingservice.entity.Shipping;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShippingRepository extends JpaRepository<Shipping, Long> {
+public interface ShippingRepository extends JpaRepository<Shipping, Integer> {
 }
